feat: reorganize and update scripts
